Processed review comments (and moved utils to one folder)
