+++ src/useLocalStore.ts
@@ -1,7 +1,7 @@
 import { observable } from "mobx"
 import { useState } from "react"
 
-import { useDeprecated } from "./utils"
+import { useDeprecated } from "./utils/utils"
 import { useAsObservableSource } from "./useAsObservableSource"
 
 export function useLocalStore<TStore extends Record<string, any>>(initializer: () => TStore): TStore
@@ -15,7 +15,7 @@ export function useLocalStore<TStore extends Record<string, any>, TSource extend
 ): TStore {
     if ("production" !== process.env.NODE_ENV)
         useDeprecated(
-            "[mobx-react-lite] 'useLocalStore' is deprecated, use 'useObservable' instead"
+            "[mobx-react-lite] 'useLocalStore' is deprecated, use 'useLocalObservable' instead."
         )
     const source = current && useAsObservableSource(current)
     return useState(() => observable(initializer(source), undefined, { autoBind: true }))[0]
