--- src/useObservable.ts
@@ -1,11 +1,0 @@
-import { observable } from "mobx"
-import React from "react"
-
-import { AnnotationsMap } from "mobx"
-
-export function useObservable<TStore extends Record<string, any>>(
-    initializer: () => TStore,
-    annotations?: AnnotationsMap<TStore, never>
-): TStore {
-    return React.useState(() => observable(initializer(), annotations, { autoBind: true }))[0]
-}
